replace old SqProject, comment some tests for now
