+++ packages/squiggle-lang/__tests__/helpers/helpers.ts
@@ -1,7 +1,7 @@
 import { run, SqValue } from "../../src/index.js";
 
 export async function testRun(x: string) {
-  const outputR = await run(x, {
+  const { output } = await run(x, {
     environment: {
       sampleCount: 1000,
       xyPointLength: 100,
@@ -9,11 +9,11 @@ export async function testRun(x: string) {
     },
   });
 
-  if (outputR.ok) {
-    return outputR.value;
+  if (output.ok) {
+    return output.value;
   } else {
     throw new Error(
-      `Expected squiggle expression to evaluate but got error: ${outputR.value}`
+      `Expected squiggle expression to evaluate but got error: ${output.value}`
     );
   }
 }
