+++ packages/squiggle-lang/__tests__/helpers/toBeOkOutput.ts
@@ -1,5 +1,3 @@
-import { SqOutputResult } from "../../src/public/types.js";
-
 declare global {
   // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace jest {
@@ -10,35 +8,35 @@ declare global {
   }
 }
 
-expect.extend({
-  toBeOkOutput(outputR: SqOutputResult, result: string, bindings: string) {
-    if (!outputR.ok) {
-      return {
-        pass: false,
-        message: () => `Expected: ok output\n` + `Received: ${outputR.value}`,
-      };
-    }
-    const gotResult = outputR.value.result.toString();
-    if (gotResult !== result) {
-      return {
-        pass: false,
-        message: () =>
-          `Expected result: ${result}\n` + `Received: ${gotResult}`,
-      };
-    }
+// expect.extend({
+//   toBeOkOutput(outputR: SqOutputResult, result: string, bindings: string) {
+//     if (!outputR.ok) {
+//       return {
+//         pass: false,
+//         message: () => `Expected: ok output\n` + `Received: ${outputR.value}`,
+//       };
+//     }
+//     const gotResult = outputR.value.result.toString();
+//     if (gotResult !== result) {
+//       return {
+//         pass: false,
+//         message: () =>
+//           `Expected result: ${result}\n` + `Received: ${gotResult}`,
+//       };
+//     }
 
-    const gotBindings = outputR.value.bindings.toString();
-    if (gotBindings !== bindings) {
-      return {
-        pass: false,
-        message: () =>
-          `Expected bindings: ${bindings}\n` + `Received: ${gotBindings}`,
-      };
-    }
+//     const gotBindings = outputR.value.bindings.toString();
+//     if (gotBindings !== bindings) {
+//       return {
+//         pass: false,
+//         message: () =>
+//           `Expected bindings: ${bindings}\n` + `Received: ${gotBindings}`,
+//       };
+//     }
 
-    return {
-      pass: true,
-      message: () => "",
-    };
-  },
-});
+//     return {
+//       pass: true,
+//       message: () => "",
+//     };
+//   },
+// });
