+++ packages/squiggle-lang/src/scripts/bench-map-reduce.ts
@@ -1,19 +1,16 @@
 #!/usr/bin/env node
 import { measure } from "../cli/utils.js";
-import { SqProject } from "../index.js";
+import { run } from "../run.js";
 
 const maxP = 5;
 
 async function main() {
   for (let p = 0; p <= maxP; p++) {
     const size = Math.pow(10, p);
-    const project = SqProject.create();
-    project.setSource(
-      "main",
-      `List.upTo(1, ${size}) -> map({|x| List.upTo(1, 100) -> reduce(0, {|a,b|a+b})})`
-    );
     const time = await measure(async () => {
-      await project.run("main");
+      await run(
+        `List.upTo(1, ${size}) -> map({|x| List.upTo(1, 100) -> reduce(0, {|a,b|a+b})})`
+      );
     });
     console.log(`1e${p}`, "\t", time);
   }
