+++ packages/squiggle-lang/src/scripts/bench-sampleset-to-pointset.ts
@@ -1,7 +1,8 @@
 #!/usr/bin/env node
 import { measure } from "../cli/utils.js";
 import { defaultEnv } from "../dists/env.js";
-import { SqProject } from "../index.js";
+import { ModuleOutput } from "../public/SqProject/ModuleOutput.js";
+import { run } from "../run.js";
 
 const maxP = 4;
 
@@ -11,27 +12,25 @@ const seed = process.env["SEED"];
 async function main() {
   for (let p = 0; p <= maxP; p++) {
     const size = Math.pow(10, p);
-    const project = SqProject.create({
-      environment: {
-        sampleCount: Number(sampleCount ?? defaultEnv.sampleCount),
-        xyPointLength: Number(sampleCount ?? defaultEnv.xyPointLength),
-        seed: seed || "default-seed",
-      },
-    });
-    project.setSource(
-      "main",
-      `
+    let result: ModuleOutput | undefined;
+    const time = await measure(async () => {
+      result = await run(
+        `
     List.upTo(1, ${size}) -> map(
       { |x| normal(x,2) -> SampleSet.fromDist -> PointSet.fromDist }
     )->List.last
-    `
-    );
-    const time = await measure(async () => {
-      await project.run("main");
+        `,
+        {
+          environment: {
+            sampleCount: Number(sampleCount ?? defaultEnv.sampleCount),
+            xyPointLength: Number(sampleCount ?? defaultEnv.xyPointLength),
+            seed: seed || "default-seed",
+          },
+        }
+      );
     });
-    const result = project.getResult("main");
-    if (!result.ok) {
-      throw new Error("Code failed: " + result.value.toString());
+    if (!result?.output.ok) {
+      throw new Error("Code failed: " + result?.output.value.toString());
     }
     console.log(`1e${p}`, "\t", time);
   }
